+++ changelog.md
@@ -3,6 +3,7 @@
 
 - - -
 ### Version 0.4.9 (Bugfix Release)
+* close #92 exclude Base64 String in log output
 * modify documentation of warning / critical times
 * add testcase.endOfStep function without warning time
 * add #81 change numbering scheme for cases and steps to always three digits to expanded color array in PHP template